[Clean up] Consolidate distribution utils.
PiperOrigin-RevId: 331359058
